ci: add -tags=assert when running tests
Loki's new query engine uses arrow-go, which uses reference counting to
know when an array can be released. While arrow-go exposes a "checked
allocator" to fail a test when not all arrays have been released, its
mechanism to assert for double-frees is hidden behind an assert build
tag.

This commit adds `-tags=assert` to running all of our tests so we can
catch double frees in the new query engine.